// plugins/src/types.ts
export type TeamId = Team['id']
// ...
export interface BasePerson {
    id: number
    team_id: number
    properties: Properties
    is_user_id: number
    is_identified: boolean
    uuid: string
    properties_last_updated_at: Record<string, any>
}

/** Raw Person row from database. */
export interface RawPerson extends BasePerson {
    created_at: string
}

/** Usable Person model. */
export interface Person extends BasePerson {
    created_at: DateTime
}
// ...
export type PropertyFilter = EventPropertyFilter | PersonPropertyFilter | ElementPropertyFilter | CohortPropertyFilter

/** Sync with posthog/frontend/src/types.ts */
export enum ActionStepUrlMatching {
    Contains = 'contains',
    Regex = 'regex',
    Exact = 'exact',
}

export interface ActionStep {
    id: number
    action_id: number
    tag_name: string | null
    text: string | null
    href: string | null
    selector: string | null
    url: string | null
    url_matching: ActionStepUrlMatching | null
    name: string | null
    event: string | null
    properties: PropertyFilter[] | null
}

/** Raw Action row from database. */
export interface RawAction {
    id: number
    team_id: TeamId
    name: string | null
    created_at: string
    created_by_id: number | null
    deleted: boolean
    post_to_slack: boolean
    slack_message_format: string
    is_calculating: boolean
    updated_at: string
    last_calculated_at: string
}

/** Usable Action model. */
export interface Action extends RawAction {
    steps: ActionStep[]
}